Create, read, update, and delete a story

Create, read, update, and delete a story

getStory(story)

createStory(...)

editStory(story, ...)

deleteStory(story)

Arguments

story
An id string or URL to a Story

...
Story attributes to either createStory or editStory. See a list of valid attributes at https://www.pivotaltracker.com/help/api/rest/v5#projects_project_id_stories_post. name is required on creation; all other fields are optional.

Value

deleteStory returns nothing, while the other functions all return a 'story' object: either the requested story (getStory), the newly created story (createStory), or the current state of the modified story editStory.

Examples

# NOT RUN {
new_bug <- createStory(
    name="Flux capacitor hangs at 0.9 gigawatts",
    description="Please investigate and fix.",
    story_type="bug"
)
new_bug <- editStory(new_bug, current_state="started")
deleteStory(new_bug)
# }